Sign-In Method and Logout Method
For my_button I just added a comma at the end of line 17.

For main.dart: changed it to show AuthPage instead of LoginPage.

Added auth_page and home_page in the pages tab.

For login_page I finished the signUserIn method and changed variables that were named "username" to  "email".

Added firebase_auth and those changes are shown in pubspec.yaml and other files.