Profile page creation
This push allows users to have their own profile page that is under the users section in the firebase realtime database. New accounts must be created to be able to use the new profile page and preexistant accounts will not be able to utilize the profile page. If a preexistant account access the profile page then the data of the app must be cleared to be able to "sign out".